fix highlight lingering issue
Previously, I use matchadd() to add highlight for a buffer. However, it
seems that matchadd() is for the current window so that when we switch
from buffer A to B, if A has trailing whitespaces, highlight will be
also shown in buffer B even if B does not have trailing whitespaces.